changed to mostly using ints for better performance

+++ embedded/include/utils/bitwise_utils.h
@@ -19,15 +19,15 @@ static inline uint32_t bv32_get_bit(const uint32_t* vector, uint32_t k) {
 
 // bit matrix operations
 
-static inline uint32_t bm32_get_bit(const BitMatrix32 matrix, uint8_t row, uint8_t col) {
+static inline uint32_t bm32_get_bit(const BitMatrix32 matrix, uint32_t row, uint32_t col) {
     return (matrix[row] >> col) & 1u;
 };
 
-static inline void bm32_set_bit(BitMatrix32 matrix, uint8_t row, uint8_t col) {
+static inline void bm32_set_bit(BitMatrix32 matrix, uint32_t row, uint32_t col) {
     matrix[row] |= 1u << col;
 };
 
-static inline void bm32_clear_bit(BitMatrix32 matrix, uint8_t row, uint8_t col) {
+static inline void bm32_clear_bit(BitMatrix32 matrix, uint32_t row, uint32_t col) {
     matrix[row] &= ~(1u << col);
 };
 
@@ -44,5 +44,5 @@ static inline uint32_t invert_bits(uint32_t x, uint32_t n) {
 }
 
 uint32_t reverse_bits(uint32_t x, uint32_t n);
-uint8_t first_set_bit(uint32_t x);
-uint8_t sum_bits(uint32_t x);
+uint32_t first_set_bit(uint32_t x);
+uint32_t sum_bits(uint32_t x);
